fixup! HAI-2162 Add link between Hankeyhteystieto and Hankekayttaja
