+++ services/hanke-service/src/main/resources/db/changelog/changesets/067-create-hankeyhteyshenkilo-table.sql
@@ -22,7 +22,7 @@ CREATE INDEX idx_hankeyhteystieto ON hankeyhteyshenkilo (hankeyhteystieto_id);
 
 COMMENT ON TABLE hankeyhteyshenkilo IS 'Table for linking the hanke users and hanke customers as contact persons.';
 COMMENT ON COLUMN hankeyhteyshenkilo.hankekayttaja_id IS 'The kayttaja who this contact is.';
-COMMENT ON COLUMN hankeyhteyshenkilo.hankeyhteystieto_id IS 'The customer this contact is represents.';
+COMMENT ON COLUMN hankeyhteyshenkilo.hankeyhteystieto_id IS 'The customer this contact represents.';
 COMMENT ON COLUMN hankeyhteyshenkilo.created_at IS 'The time this row was originally created.';
 
 ALTER TABLE hankeyhteystieto
